fix: Smooth scrolling when clicking on the navbar now works.
