+++ src/app/page.tsx
@@ -9,31 +9,9 @@ import ContactSection from '../components/ContactSection';
 import Footer from '@/components/Footer';
 
 const Page: React.FC = () => {
-  const [isNavbarVisible, setIsNavbarVisible] = useState(true);
-  const [lastScrollY, setLastScrollY] = useState(0);
-
-  const handleScroll = () => {
-    const currentScrollY = window.scrollY;
-    if (currentScrollY > lastScrollY) {
-      // Scrolling down
-      setIsNavbarVisible(false);
-    } else {
-      // Scrolling up
-      setIsNavbarVisible(true);
-    }
-    setLastScrollY(currentScrollY);
-  };
-
-  useEffect(() => {
-    window.addEventListener('scroll', handleScroll);
-    return () => {
-      window.removeEventListener('scroll', handleScroll);
-    };
-  }, [lastScrollY]);
-
   return (
     <main className="flex min-h-screen flex-col p-24 bg-background">
-      <Navbar isVisible={isNavbarVisible} />
+      <Navbar />
       <HomeSection />
       <ProjectSection />
       <AboutSection />
@@ -41,7 +19,6 @@ const Page: React.FC = () => {
       <Footer />
     </main>
   );
-}
-
+};
 
 export default Page;
