+++ src/components/Navbar.tsx
@@ -1,28 +1,59 @@
-import React, { useState, useEffect } from "react";
+import React, { useState, useEffect } from 'react';
 
-interface NavbarProps {
-    isVisible: boolean;
-}
+const Navbar: React.FC = () => {
+    const [isNavbarVisible, setIsNavbarVisible] = useState(true);
+    const [lastScrollY, setLastScrollY] = useState(0);
 
-const Navbar: React.FC<NavbarProps> = ({ isVisible }) => {
+    useEffect(() => {
+        const handleScroll = () => {
+            const currentScrollY = window.scrollY;
+            if (currentScrollY > lastScrollY) {
+                // Scrolling down
+                setIsNavbarVisible(false);
+            } else {
+                // Scrolling up
+                setIsNavbarVisible(true);
+            }
+            setLastScrollY(currentScrollY);
+        };
+
+        window.addEventListener('scroll', handleScroll);
+
+        return () => {
+            window.removeEventListener('scroll', handleScroll);
+        };
+    }, [lastScrollY]);
+
     const scrollToTop = () => {
         window.scrollTo({
             top: 0,
-            behavior: "smooth"
+            behavior: 'smooth'
         });
     };
 
+    const scrollToSection = (sectionId: string) => {
+        const section = document.getElementById(sectionId);
+        if (section) {
+            window.scrollTo({
+                top: section.offsetTop,
+                behavior: 'smooth'
+            });
+        }
+    };
+
     return (
-        <nav className={`fixed top-0 left-0 right-0 z-20 bg-background transition-transform duration-300 ${isVisible ? 'translate-y-0' : '-translate-y-full'}`}>
+        <nav className={`fixed top-0 left-0 right-0 z-20 bg-background transition-transform duration-300 ${isNavbarVisible ? 'translate-y-0' : '-translate-y-full'}`}>
             <div className="container mx-auto flex justify-between items-center py-4 text-lg">
                 <p className="text-primary hover:underline text-darkorange" onClick={scrollToTop}>
                     Tilde Eriksen Eine
                 </p>
-
                 <ul className="flex text-darkblue">
                     {['Projects', 'About me', 'Contact'].map((item) => (
                         <li key={item} className="mr-6">
-                            <a href={`#${item.toLowerCase().replace(' ', '-')}`} className="text-primary hover:underline">
+                            <a
+                                className="text-primary hover:underline"
+                                onClick={() => scrollToSection(item.toLowerCase().replace(' ', '-'))}
+                            >
                                 {item}
                             </a>
                         </li>
@@ -31,6 +62,6 @@ const Navbar: React.FC<NavbarProps> = ({ isVisible }) => {
             </div>
         </nav>
     );
-}
+};
 
 export default Navbar;
